refactor: extract register on app to file
